Cleaning up some unit tests
