+++ library/src/test/java/com/corycharlton/bittrexapi/internal/gson/GsonFactoryTest.java
@@ -1,0 +1,19 @@
+package com.corycharlton.bittrexapi.internal.gson;
+
+import org.junit.Test;
+import org.junit.experimental.runners.Enclosed;
+import org.junit.runner.RunWith;
+
+import static org.junit.Assert.*;
+
+@RunWith(Enclosed.class)
+public class GsonFactoryTest {
+
+    public static class When_getGson_is_called {
+
+        @Test
+        public void it_should_return_an_instance_of_Gson() {
+            assertNotNull(GsonFactory.getGson());
+        }
+    }
+}
